+++ src/components/Navbar/index.js
@@ -3,44 +3,37 @@ import { connect } from 'react-redux';
 
 // Material-UI Core
 import { withStyles } from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import InputBase from '@material-ui/core/InputBase';
-import IconButton from '@material-ui/core/IconButton';
-import Badge from '@material-ui/core/Badge';
-import SwipeableDrawer from '@material-ui/core/SwipeableDrawer';
-import Menu from '@material-ui/core/Menu';
-import MenuItem from '@material-ui/core/MenuItem';
-import Avatar from '@material-ui/core/Avatar';
-import Typography from '@material-ui/core/Typography';
-import Button from '@material-ui/core/Button';
+import {
+  AppBar,
+  IconButton,
+  Toolbar,
+  SwipeableDrawer,
+} from '@material-ui/core';
 
 // Material-UI Icons
 import MenuIcon from '@material-ui/icons/Menu';
-import SearchIcon from '@material-ui/icons/Search';
-import ShoppingCartIcon from '@material-ui/icons/ShoppingCartOutlined';
-import AccountCircle from '@material-ui/icons/AccountCircle';
 
+// React Components
 import Logo from './Logo';
-import styles from './styles';
+import Search from './Search';
 import Sidelist from './Sidelist';
+import Cart from './Cart';
+import Login from './Login';
+import Profile from './Profile';
+
+import styles from './styles';
 
 class Navbar extends Component {
   state = {
     drawerOpened: false,
-    profileOpened: false,
   };
 
   toggleDrawer = open => {
     this.setState({ drawerOpened: open });
   };
 
-  toggleProfile = open => {
-    this.setState({ profileOpened: open });
-  };
-
   render() {
-    const { classes, itemsInCart } = this.props;
+    const { classes, user } = this.props;
     return (
       <div className={classes.root}>
         <AppBar
@@ -83,91 +76,14 @@ class Navbar extends Component {
               <MenuIcon />
             </IconButton>
             <Logo classes={classes} />
-            <div className={classes.search}>
-              <div className={classes.searchIcon}>
-                <SearchIcon style={{ color: 'grey' }} />
-              </div>
-              <InputBase
-                placeholder="Search…"
-                classes={{
-                  root: classes.inputRoot,
-                  input: classes.inputInput,
-                }}
-              />
-            </div>
-            <IconButton className={classes.IconButton}>
-              <Badge badgeContent={itemsInCart} color="secondary">
-                <a
-                  href="#/cart"
-                  style={{ textDecoration: 'none', color: 'grey' }}
-                >
-                  <ShoppingCartIcon />
-                </a>
-              </Badge>
-            </IconButton>
+            <Search classes={classes} />
+            <Cart classes={classes} />
             <div className={classes.sectionDesktop}>
-              <a href="#/login">
-                <IconButton
-                  className={classes.IconButton}
-                  aria-owns={open ? 'menu-appbar' : undefined}
-                  aria-haspopup="true"
-                  onClick={() => {
-                    this.toggleProfile(!this.state.profileOpened);
-                  }}
-                >
-                  <AccountCircle />
-                </IconButton>
-              </a>
-              {/* <div>
-                <Menu
-                  id="menu-appbar"
-                  anchorOrigin={{
-                    vertical: 'top',
-                    horizontal: 'right',
-                  }}
-                  transformOrigin={{
-                    vertical: 'top',
-                    horizontal: 'right',
-                  }}
-                  open={this.state.profileOpened}
-                  onClose={() => {
-                    this.toggleProfile(false);
-                  }}
-                  style={{ top: '50px' }}
-                >
-                  <MenuItem style={{ height: '100px' }}>
-                    <Avatar
-                      alt="Remy Sharp"
-                      src="https://material-ui.com/static/images/avatar/1.jpg"
-                      style={{ margin: 10, width: 60, height: 60 }}
-                    />
-                    <div
-                      style={{
-                        display: 'flex',
-                        flexDirection: 'column',
-                        justifyContent: 'center',
-                        marginLeft: '1rem',
-                      }}
-                    >
-                      <Typography variant="subtitle2">Foo Bar</Typography>
-                      <Typography variant="caption">
-                        fooBar123@gmail.com
-                      </Typography>
-                      <Button
-                        variant="contained"
-                        color="secondary"
-                        size="small"
-                        className={classes.button}
-                        onClick={() => {
-                          console.log('clicked');
-                        }}
-                      >
-                        Logout
-                      </Button>
-                    </div>
-                  </MenuItem>
-                </Menu>
-              </div> */}
+              {user.imageUrl ? (
+                <Profile user={user} />
+              ) : (
+                <Login classes={classes} />
+              )}
             </div>
           </Toolbar>
         </AppBar>
@@ -176,13 +92,8 @@ class Navbar extends Component {
   }
 }
 
-const mapStateToProps = ({ lineItems = [] }) => {
-  const itemsInCart = lineItems.reduce((sum, { quantity }) => {
-    sum += quantity;
-    return sum;
-  }, 0);
-
-  return { itemsInCart };
-};
+const mapStateToProps = ({ user }) => ({
+  user,
+});
 
 export default connect(mapStateToProps)(withStyles(styles)(Navbar));
